+++ tests/testthat/test-html-builders.R
@@ -28,10 +28,8 @@ test_that("HTML applies table and row break rules", {
   breakable(ht) <- TRUE
   html <- to_html(ht)
   expect_match(html, "break-inside: auto; page-break-inside: auto;", fixed = TRUE)
-  # we do this because print_html() prints some css via a private function
-  printed <- paste(capture.output(print_html(ht)), collapse = "\n")
   expect_match(
-    printed,
+    huxtable_html_css(),
     ".huxtable tr {\n  break-inside: avoid;\n  page-break-inside: avoid;",
     fixed = TRUE
   )